requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
crx-unpack==0.1.4
discord.py==1.0.0a0
docopt==0.6.2
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.5.2
Pillow==6.0.0
pycodestyle==2.5.0
pylint==2.3.1
six==1.12.0
typed-ast==1.3.1
virtualenv==16.7.3
websockets==6.0
wrapt==1.11.1
yarl==1.3.0
